pyproject.toml
[project]
name = "reminder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "passlib[bcrypt]>=1.7",
    "python-dateutil>=2.9",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "PyJWT>=2.9",
    "tzdata>=2024.1",
    "email-validator>=2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "testcontainers[postgres]>=4.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
